requirements.txt
inquirer==2.10.0
mysql-connector-python==8.0.31
pandas==1.5.0